Improved uninstall script.

+++ uninstall.sh
@@ -1,9 +1,20 @@
 #!/usr/bin/env bash
 
+mccli_path="$(which mccli)"
+
 set -e
 
-unlink "$(which mccli)"
+if [ "$mccli_path" = "" ] && [ ! -e "$(dirname "$(readlink -f "$0")")/mccli.sh" ]; then
+	echo "mccli: cannot uninstall, mccli is not installed, exiting";
+	exit 1;
+fi
 
-rm "$(dirname "$(readlink -f "$0")")/mccli.sh"
+if [ ! "$mccli_path" = "" ]; then
+	unlink "$mccli_path"
+fi
+
+if [ -e "$(dirname "$(readlink -f "$0")")/mccli.sh" ]; then
+	rm "$(dirname "$(readlink -f "$0")")/mccli.sh"
+fi
 
 echo "mccli: sucessfully uninstalled"
